% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murlib.data-package.R
\docType{package}
\name{murlib.data-package}
\alias{murlib.data}
\alias{murlib.data-package}
\title{murlib.data: Data Utilities}
\description{
A collection of utilities to solve common data problems.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
